t1/backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database & Cache
pymongo==4.6.0
redis==5.0.1
motor==3.3.2

# Security & Authentication
pyjwt==2.8.0
cryptography==46.0.6
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4

# HTTP & API
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Data Processing
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.1.0

# Task Queue & Async
celery==5.3.4
kombu==5.3.5

# Web Automation
playwright==1.39.0
selenium==4.15.2

# LLM Integration
openai==1.3.7  # For GLM API compatibility
anthropic==0.7.8

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
tenacity==8.2.3

# Monitoring & Logging
loguru==0.7.2
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# Rate Limiting
slowapi==0.1.9

# CORS
python-jose[cryptography]==3.3.0
